refactor(mobile): use a composite component for the bookmark card
